Allow missing and null pointer fields to be nil.
This commit ensures missing fields for pointer field types remain nil:

  * This allows one to distinguish the field wasn't present from the
    field being NULL when the PtrNil option is false.
  * This also makes it more convenient to reuse structs with pointer
    field types with various encoders like the standard library JSON
    marshaler, where you may want to omit a field if it didn't exist in
    the queried results.

The expectations around this change are captured in the
TestMissingFields test.

In addition, it also fixes an edge case where even with PtrNil set to
true, pointer field types would always return with a pointer to a zero
value instead of being nil even when the query property's value is
VT_NULL.

This fix is checked by the TestNullPointerField test.